Refactor: Add custom queries mod
